better similarity structure, more generic Float (currently f32) instead of f64, most of grid_search_optimization implemented

+++ myrio-core/src/lib.rs
@@ -11,6 +11,6 @@ pub mod utils;
 #[macro_export]
 macro_rules! assert_float_eq {
     ($lhs: expr, $rhs: expr) => {
-        assert!(($lhs - $rhs).abs() < f64::EPSILON)
+        assert!(($lhs - $rhs).abs() < f32::EPSILON)
     };
 }
